feat: block all external reqs when account is not depolyed
