+++ apps/extension/src/entries/side-panel/pages/Alert.tsx
@@ -1,32 +1,29 @@
 import { useApproval } from '../contexts/approval-context';
 import { ethErrors } from 'eth-rpc-errors';
 import Spin from '@/components/Spin';
-import { useEffect, useState } from 'react';
 import { Card, CardContent, CardFooter } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
-import { UnSupportedMethod } from '@/background/provider/rpcFlow/checkMethodExist';
+import { UnsupportedMethod } from '@/background/provider/rpcFlow/checkCallable';
 
-export default function Connect() {
+export default function BlockedAlert() {
   const { approval, reject } = useApproval();
-  const [loading, setLoading] = useState(true);
-  useEffect(() => {
-    if (approval?.data?.options) {
-      setLoading(false);
-    }
-  }, [approval]);
+
   if (!approval || !approval.data) {
-    return <Spin isLoading={loading} />;
+    return <Spin isLoading={true} />;
   }
 
   const {
     data: { dApp, options },
   } = approval;
 
+  const { name, reason } = options as UnsupportedMethod;
+
   const handleReject = () => {
     reject(ethErrors.provider.userRejectedRequest());
     window.close();
   };
+
   return (
     <div className="w-full h-full flex justify-center items-center min-h-screen p-4">
       <Card className="w-full h-full flex flex-col justify-between border-none rounded-none shadow-none">
@@ -38,14 +35,12 @@ export default function Connect() {
             </Avatar>
           </div>
           <div className="mb-6">
-            <div className="text-3xl font-medium mb-2">{dApp.name}</div>
-            <div className="text-gray-300 text-lg font-normal">
-              {dApp.origin}
-            </div>
+            <div className="elytro-text-subtitle">{dApp.name}</div>
+            <div className="elytro-text-body">{dApp.origin}</div>
           </div>
-          <div className="text-lg font-medium">
-            The Dapp called {(options as UnSupportedMethod).name} failed, the
-            reason is {(options as UnSupportedMethod).reason}
+          <div className="elytro-text-body">
+            The Dapp called <span className="font-bold">{name}</span> failed,
+            the reason is <span className="font-bold">{reason}</span>
           </div>
         </CardContent>
         <CardFooter className="flex justify-between space-x-4">
